requirements.txt
paho-mqtt
logzero  # see github.com/metachris/logzero